Exercise 4.21

+++ controllers/blog.js
@@ -2,6 +2,7 @@ const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 const jwt = require("jsonwebtoken");
+const e = require("express");
 
 blogRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("users");
@@ -9,6 +10,8 @@ blogRouter.get("/", async (request, response) => {
 });
 
 blogRouter.post("/", async (request, response, next) => {
+  // Token verification stuff
+
   const token = request.token;
   if (!token) {
     return response.status(401).json({ error: "token missing" });
@@ -26,6 +29,7 @@ blogRouter.post("/", async (request, response, next) => {
     return response.status(401).json({ error: "token invalid" });
   }
 
+  // Actual create stuff
   const user = await User.findById(decodedToken.id);
 
   const blogObject = {
@@ -50,9 +54,36 @@ blogRouter.post("/", async (request, response, next) => {
 });
 
 blogRouter.delete("/:id", async (request, response, next) => {
+  // Token verification stuff
+  const token = request.token;
+  if (!token) {
+    return response.status(401).json({ error: "token missing" });
+  }
+
+  // verify can throw; handle here
+  let decodedToken;
   try {
-    const blogs = await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    decodedToken = jwt.verify(token, process.env.SECRET);
+  } catch (_error) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const userid = decodedToken.id;
+
+  // Actual delete stuff
+  try {
+    const blog = await Blog.findById(request.params.id);
+    if (blog === null) {
+      response.status(404).end();
+    } else if (blog.user.toString() === userid.toString()) {
+      const blogs = await Blog.findByIdAndRemove(request.params.id);
+      response.status(204).end();
+    } else {
+      response.status(403).end();
+    }
   } catch (error) {
     next(error);
   }
